==> notebooks/ETL/Google/2.categorias.ipynb <==
### Este Notebook estara enfocado en el desanidado y normalizacion de los datos de la columna `category`. para filtrar el dataset original

import pandas as pd

# Leemos la tabla g_sitios que tenemos hasta el momento.
g_sitios = pd.read_parquet(r'C:\Users\mauri\OneDrive\Escritorio\proyectogrupal\data\google_sitios.parquet', engine='pyarrow')

# Creamos una nueva tabla de categorias y desempaquetamos su información.
categorias = g_sitios[['gmap_id', 'category']].explode('category')

# Borramos las columnas `MISC`, `relative_results` y `category`
g_sitios = g_sitios.drop(columns=['MISC', 'relative_results', 'category'])

# Guaradamos la tabla desanidada que se conecta a travez de gmap_id con la tabla principal.
categorias.to_parquet('datasets/google/sitios/categories_sitios_google_todavia_nombres.parquet')

categorias = pd.read_parquet(r'C:\Users\mauri\OneDrive\Escritorio\proyectogrupal\data\dataset_gastronomico_google.parquet', engine='pyarrow')
# Obtener lista única de categorías en 'category'
categorias_unicas = categorias['category'].unique()
# Mostrar la lista
print(categorias_unicas)

# Filtrar las categorias por una categoria general para un id_unico
categorias.loc[:, 'category'] = categorias['category'].str.lower()
categorias = categorias[~categorias['category'].str.contains('night club|buffet|school|food products supplier|court|banquet hall|supplier|thermal baths|spa|class', case=False, na=False)]

categorias.loc[:, 'category_general'] = None

categorias.loc[categorias['category'].str.contains('bar|shochu|liquor|beer|pub|cider|drink|wine|beverage|bodega|tavern|vineyard|sake', case=False, na=False), 'category_general'] = 'drinks'
categorias.loc[categorias['category'].str.contains('cafe|coffee|te|cafeteria|tea|patisserie|confectionery|lounge', case=False, na=False), 'category_general'] = 'coffee and te'
categorias.loc[categorias['category'].str.contains('restaurant|carvery|brasserie|bistro|restaurante|grill|food|bbq|pizza|soup|crab|crêperie', case=False, na=False), 'category_general'] = 'restaurant'
categorias.loc[categorias['category'].str.contains('shop|deli|frituur|greengrocer|fruit|hawker stall|wholesaler|store|supermarket|bakery|takeaway|delivery|service|stand|churreria|fast food|takeout|sweets|market', case=False, na=False), 'category_general'] = 'shops and takeaway'

# Crear un diccionario de categorías generales y sus IDs únicos
category_ids = {
    'drinks': 1,
    'coffee and te': 2,
    'restaurant': 3,
    'shops and takeaway': 4
}

# Asignar el ID único correspondiente a cada categoría general
categorias['category_general_id'] = categorias['category_general'].map(category_ids)
# Eliminar duplicados basados en la columna 'gmap_id' en el DataFrame 'categorias'
categorias_unicas = categorias.drop_duplicates(subset='gmap_id', keep='first')
categorias_unicas

# Unimos los id con la tabla principal.
g_sitios= pd.merge(g_sitios, categorias_unicas[['gmap_id', 'category_general_id']], on='gmap_id', how='left')

# Eliminar filas donde 'category_id' sea None o NaN
g_sitios = g_sitios[g_sitios['category_general_id'].notna()]

# Convertir la columna 'category_id' a tipo int
g_sitios['category_general_id'] = g_sitios['category_general_id'].astype(int)

g_sitios.head()

g_sitios.to_parquet('v0_g_sitios.parquet')

categorias

categorias.to_parquet('v0_categorias.parquet')
