==> notebooks/ETL/Google/3.address.ipynb <==
### Este Notebook esta enfocado en como conectar los estados con las reviews y Yelp

La columna `address` de la tabla principal `g_sitios` contiene informacion importante sobre como conectar las reviews de los estados y los negocios.

si se normaliza esta columna en una tabla aparte que contenga: la ciudad, la direccion, y codigo postal, mas la clave unica gmap_id, podemos agrupar por estados y crear un identificador unico.

con un eda sobre las categorias de los negocios en cada estado, podremos decidir que estados usar y cuales no, basandonos en datos y no en suposiciones.

import pandas as pd
import os

# Leemos la tabla g_sitios que tenemos hasta el momento.
g_sitios = pd.read_parquet(r'C:\Users\mauri\OneDrive\Escritorio\proyectogrupal\data\v0_g_sitios.parquet', engine='pyarrow')
g_sitios.info()

# Creamos la nueva tabla address que se conectara mediante gmap_id con la tabla principal.
address = g_sitios[['gmap_id','address', 'name']]
address.head()

# Separar la columna 'address' en nuevas columnas
address_split = address['address'].str.split(',', n=3, expand=True)

# Cambiar el nombre de las columnas
address_split.columns = ['nombre', 'direccion', 'ciudad', 'cod.postal']

# Agregar la identificacion unica a las nuevas columnas separadas.
address = address[['gmap_id','address']].join(address_split)
address.head()

# Agregar una columna nueva llamada estados.
address['estado'] = None

# Listas de los 51 códigos postales y nombres de estados.
codigos_postales = [
    'CA', 'NY', 'IA', 'GA', 'FL', 'TX', 'LA', 'OR', 'WV', 'VA', 
    'AR', 'PA', 'NM', 'NC', 'TN', 'WI', 'NJ', 'IN', 'IL', 'DC', 
    'MD', 'ME', 'NE', 'WA', 'MI', 'OH', 'OK', 'MO', 'KS', 'UT', 
    'HI', 'NV', 'AZ', 'AL', 'CO', 'MA', 'ID', 'SC', 'RI', 'KY', 
    'AK', 'MT', 'MN', 'CT', 'MS', 'SD', 'WY', 'NH', 'DE', 'VT', 
    'ND'
]

nombres_estados = [
    'California', 'New York', 'Iowa', 'Georgia', 'Florida', 'Texas', 
    'Louisiana', 'Oregon', 'West Virginia', 'Virginia', 'Arkansas', 
    'Pennsylvania', 'New Mexico', 'North Carolina', 'Tennessee', 'Wisconsin', 
    'New Jersey', 'Indiana', 'Illinois', 'District of Columbia', 'Maryland', 
    'Maine', 'Nebraska', 'Washington', 'Michigan', 'Ohio', 'Oklahoma', 
    'Missouri', 'Kansas', 'Utah', 'Hawaii', 'Nevada', 'Arizona', 
    'Alabama', 'Colorado', 'Massachusetts', 'Idaho', 'South Carolina', 
    'Rhode Island', 'Kentucky', 'Alaska', 'Montana', 'Minnesota', 
    'Connecticut', 'Mississippi', 'South Dakota', 'Wyoming', 
    'New Hampshire', 'Delaware', 'Vermont', 'North Dakota'
]

# cod.postal contiene las letras del estado, podemos filtrar los estados rapidamente, a travez de un bucle entre codigo y nombres.
for codigo, estado in zip(codigos_postales, nombres_estados):
    address.loc[address['cod.postal'].str.contains(codigo, na=False), 'estado'] = estado

# El cod.postal puede ser nulo por falta de informacion ya que se corre la informacion hacia la izquierda.
# Para solucionar esto filtramos los cod.postales nulos que se corrieron hacia la izquierda
for codigo, estado in zip(codigos_postales, nombres_estados):
    address.loc[(address['cod.postal'].isnull()) & (address['ciudad'].str.contains(codigo, na=False)), 'estado'] = estado

# Ahora que estan corregidos las mayorias de los cod.postal algunos datos se cargaron al reves, conteniendo el cod en nombre y el nombre en cod.postal.
for estado in nombres_estados:
    address.loc[(address['estado'].isnull()) & (address['nombre'].str.contains(estado, na=False)), 'estado'] = estado

# Eliminar filas donde al menos 5 columnas son nulas
columnas = ['gmap_id', 'nombre', 'direccion', 'ciudad', 'cod.postal', 'estado']
address = address.dropna(thresh=len(columnas) - 1)

# Ultimas condicionales: el cod.postal no es nulo, pero su codigo aun se encuentra en la ciudad.
for codigo, estado in zip(codigos_postales, nombres_estados):
    address.loc[(address['cod.postal'].notnull()) & (address['ciudad'].str.contains(codigo, na=False)), 'estado'] = estado

# El cod.postal no es nulo, pero esta escrito el estado en la ciudad.  
for codigo, estado in zip(codigos_postales, nombres_estados):
    address.loc[(address['cod.postal'].notnull()) & (address['ciudad'].str.contains(estado, na=False)), 'estado'] = estado
    
# El cod.postal contiene escrito el estado y no un codigo.  
for codigo, estado in zip(codigos_postales, nombres_estados):
    address.loc[(address['cod.postal'].notnull()) & (address['cod.postal'].str.contains(estado, na=False)), 'estado'] = estado

# Terminada la normalizacion solo quedaron muy pocas filas donde no se puede saber su estado. Se eliminaran.
address = address.dropna(subset=['estado'])
address = address.drop(columns=['nombre'])

# Agregamos un identificador unico a cada estado para que se conecte con la tabla principal g_sitios.
address['id_estado'] = address['estado'].factorize()[0] + 1
address['id_estado'] = address['id_estado'].astype('int32')

# Mezclamos la informacion con la tabla principal.
g_sitios = g_sitios.merge(address[['gmap_id', 'id_estado']], on='gmap_id', how='left')
g_sitios = g_sitios.drop(columns=['address'])
g_sitios = g_sitios.drop(columns=['state']) # Informacion estatica. No sirve

# Borrar filas con id_estado nulos, ya que no van a tener informacion para conectarse con reviews sin un estado.
g_sitios = g_sitios.dropna(subset=['id_estado'])
g_sitios['id_estado'] = g_sitios['id_estado'].astype('int32') # tipo de dato entero.

# Optimizando tipo de datos.
g_sitios['latitude'] = g_sitios['latitude'].astype('float32')
g_sitios['longitude'] = g_sitios['longitude'].astype('float32')
#  Extraemos las columnas latitud y longitud de g_sitios.
lat_long = g_sitios[['gmap_id', 'latitude', 'longitude']]

# Merge de lat_long a address usando id_negocio como clave.
address = address.merge(lat_long[['latitude','longitude','gmap_id']], on='gmap_id', how='left')
# Eliminar las columnas latitud y longitud de g_sitios
g_sitios = g_sitios.drop(columns=['latitude', 'longitude'])

# Contamos la cantidad de restaurantes y hoteles que contienen los estados, para elegir un Top 5 de mayor cantidad de negocios 
conteo_negocios_por_estado = g_sitios.groupby('id_estado').size().reset_index(name='cantidad_negocios')
conteo_negocios_por_estado.sort_values(by='cantidad_negocios', ascending=False).head(5)

# Filtrar las filas donde 'id_estado' sea 24 o 9
filtered_address = address[address['id_estado'].isin([9,24,18,7,13])]

# Mostrar solo las columnas 'estado' e 'id_estado'
filtered_address = filtered_address[['estado', 'id_estado']]
filtered_address.value_counts()

google_estados = [9,24,18,7,13]
g_sitios = g_sitios[g_sitios['id_estado'].isin(google_estados)]
g_sitios = g_sitios.drop(columns=['price','description'])
g_sitios.reset_index()
g_sitios.info()

address = pd.read_parquet(r'C:\Users\mauri\OneDrive\Escritorio\proyectogrupal\data\v0_address.parquet', engine='pyarrow')

address['id_estado'] = address['id_estado'].astype(int)
# Terminado la informacion. guardamos el dataframe en parquet.
os.makedirs('../../../data', exist_ok=True)
address.to_parquet('../../../data/v0_address.parquet', index=False)

# Muestra parcial de la tabla:
address.head()

# Muestra parcial de la tabla:
address.info()

# Terminado la informacion. guardamos el dataframe principal actualizado en parquet.
os.makedirs('../../../data', exist_ok=True)
g_sitios.to_parquet('../../../data/v1.g_sitios.parquet', index=False)

# Muestra parcial de la tabla:
g_sitios.head()

# Muestra parcial de la tabla:
g_sitios.info()

Importare los dataset para cambiar el nombre de las columnas llamadas "gmap_id" por "id_negocio".

import pandas as pd
g_sitios = pd.read_parquet(r"C:\Users\mauri\OneDrive\Escritorio\proyectogrupal\data\v1.g_sitios.parquet", engine='pyarrow')
g_sitios

address = pd.read_parquet(r"C:\Users\mauri\OneDrive\Escritorio\proyectogrupal\data\v0_address.parquet", engine='pyarrow')
address

address = address.rename(columns={'gmap_id': 'id_negocio'})
g_sitios = g_sitios.rename(columns={'gmap_id': 'id_negocio'})

address.to_parquet("v0_address.parquet")
g_sitios.to_parquet("v1_g_sitios.parquet")

address = address.rename(columns={'cod.postal': 'cod_postal'})
address.to_parquet("v0_address.parquet")


import pandas as pd
address = pd.read_parquet(r"C:\Users\mauri\OneDrive\Escritorio\proyectogrupal\data\v0_address.parquet", engine='pyarrow')
address.info()


# Supongamos que tienes un DataFrame llamado df
# Supongamos que tu DataFrame se llama df

# Localiza las filas donde el 'Address' sea igual a 'Bubba\'s Texas Cookin\', Montana 59912'
address.loc[address['address'] == "Bubba's Texas Cookin', Montana 59912", 'ciudad'] = 'Columbia Falls'
address.loc[address['address'] == "Bubba's Texas Cookin', Montana 59912", 'cod_postal'] = 'MT 59912'

address.loc[address['address'] == "New York, 美国明记小炒邮政编码: 10002", 'address'] = '237 Grand St'
address.loc[address['id_negocio'] == "0x89c259e473970ed3:0xa76c925ac50da6c6", 'ciudad'] = 'New York'
address.loc[address['id_negocio'] == "0x89c259e473970ed3:0xa76c925ac50da6c6", 'cod_postal'] = 'NY 10002'


address['cod_postal'] = address['cod_postal'].fillna(address['ciudad'])


address.to_parquet("v1_address.parquet")
