==> notebooks/ETL/Google/4.misc.ipynb <==
### Este Notebook estara enfocado en el desanidado y normalizacion de los datos de la columna `MISC`. Al terminar continuar con el Notebook 5.

import pandas as pd

#  Abrimos el dataframe que contiene toda la data de los negocios de google.
g_sitios = pd.read_parquet(r"C:\Users\Juli\Desktop\Data Science\Archivos Proyecto Final\sitios_google_columnas_originales.parquet")

g_sitios

# Creamos una nueva tabla misc
misc_sitios = g_sitios[['gmap_id', 'MISC']]

# Donde MISC sea NULO no hay Información Se pueden Borrar ya que esta tabla es SOLO MISCELANEOS.
misc_sitios = misc_sitios.dropna(subset=['MISC'])

# Se Define una Función que convierte los Diccionarios en Listas para poder Aplicar `explode` despues.  Sólo se toman aquellos pares del diccionario donde value is not None
def dict_to_list(diccionario:dict) -> list:
    """ Devuelve el Diccionario de la columna `MISC` y lo retorna
                    como una lista de String "key:Value" Siempre y cuando el 
                    value de la pareja no sea Nulo.

    Args:
        diccionario (dict): diccionario recibido

    Returns:
        list: Lista a devolver
    """
    return [f"{key}: {value}" for key, value in diccionario.items() if value is not None]

# Aplicamos la funcion definida arriba.
misc_sitios['lista_misc'] = misc_sitios['MISC'].apply(dict_to_list)

# Muestra parcial de la tabla:
misc_sitios.head()

Se borra la columna MISC, se aplica explode al dataframe con respecto a la columna `lista_misc` y luego renombrar la columna para tener lista la tabla

misc_sitios = misc_sitios.drop(columns=['MISC'])

misc_sitios = misc_sitios.explode('lista_misc')

misc_sitios.rename(columns={'lista_misc': 'misc'}, inplace=True)
misc_sitios.head()

# Guaradamos la tabla desanidada que se conecta a travez de gmap_id con la tabla principal.
misc_sitios.to_parquet(r"C:\Users\Juli\Desktop\Data Science\Archivos Proyecto Final\google_misc_sitios.parquet")

Filtrare la tabla con los estados y categorías seleccionadas para este proyecto. Para ello, importaré el dataframe "vi.g_sitios". Realizaré una unión entre ambos dataframes para luego filtrar por las categorías y estados elegidos. 

sitios = pd.read_parquet(r"C:\Users\Juli\Desktop\Data Science\Archivos Proyecto Final\v1.g_sitios.parquet")
sitios.head()

df_merged = pd.merge(misc_sitios, sitios[['gmap_id']], on='gmap_id', how='inner')
df_merged

pd.set_option('display.max_colwidth', None)
df_merged

# Extraer el texto antes del primer ':'
df_merged['category'] = df_merged['misc'].str.split(':').str[0]

# Obtener las categorías únicas
unique_categories = df_merged['category'].unique()
df_merged

# Extraer el contenido después de ':' en la columna 'misc'
df_merged['misc_content'] = df_merged['misc'].str.split(':').str[1].str.strip()
# Usar una expresión regular para extraer cada texto dentro de comillas simples
df_merged['misc_content'] = df_merged['misc'].str.extractall(r"'(.*?)'").groupby(level=0).agg(list)
# Separar la lista en filas
df_merged = df_merged.explode('misc_content')
# Muestra de la tabla
df_merged = df_merged[['gmap_id', 'category', 'misc_content']]

df_merged.rename(columns={'category': 'categoria_atributo'}, inplace=True)
df_merged.rename(columns={'misc_content': 'atributo'}, inplace=True)
df_merged.rename(columns={'gmap_id': 'id_negocio'}, inplace=True)
df_merged.info()

df_merged

Conteo de nulos

nulos_por_columna = df_merged.isnull().sum()
nulos_por_columna

# Eliminar filas nulas
df_merged = df_merged[df_merged['atributo'].notnull()]
df_merged.info()

Veamos cuantas categorias de atributos hay

categorias = df_merged['categoria_atributo'].unique()
print(categorias,'\ntotal categorias:',len(categorias))

### Existe una categoria duplicada de distinto nombre: 
- `Health & safety`
- `Health and safety`
La agrupamos en una sola.

df_merged.loc[:, 'categoria_atributo'] = df_merged['categoria_atributo'].replace({'Health & safety': 'Health and safety'})

veamos si hay relaciones en otras categorias para agruparlas

categorias = df_merged['categoria_atributo'].unique()

# Iterar a través de cada categoría y mostrar los atributos únicos
for categoria in categorias:
    atributos = df_merged[df_merged['categoria_atributo'] == categoria]['atributo'].unique()
    print(f"{categoria}\n", atributos, "\n")

### Viendo los atributos comprendemos que Health and Safety es un atributo de la pandemia el cual hoy ya no es necesario ni tenido en cuenta para lugares turisticos. Por ende vamos a eliminar ese atributo.

# Quitamos Health and Safety de los atributos.
df_merged = df_merged[df_merged['categoria_atributo'] != 'Health and safety']
# Comprobamos
categorias = df_merged['categoria_atributo'].unique()
print(categorias)

### Tambien vemos muchos atributos repetidos. Algunos contienen - y otros solo espacios. Asi que reemplazamos - por espacios y ver de nuevo las categorias

Ademas generalizar todos los tipos de inclusion sexual en una sola categoria. 'LGBTQ Friendly'

# Quitar el guion en los valores de 'atributo'
df_merged.loc[:, 'atributo'] = df_merged['atributo'].str.replace('-', ' ', regex=False)

# Generalizar el atributo.
mapeo_lgbtq = {
    'LGBTQ friendly': 'LGBTQ friendly',
    'LGBTQ-friendly': 'LGBTQ friendly',
    'Transgender safespace': 'LGBTQ friendly',
    'Transgender safe space': 'LGBTQ friendly'
}

mapeo_accesibilidad = {
    'Wheelchair accessible restroom': 'Wheelchair accessible toilet',
    'Wheelchair accessible parking lot': 'Wheelchair accessible car park',
    'Wheelchair accessible lift': 'Wheelchair accessible elevator'
}

# Aplicar el mapeo a la columna 'atributo'
df_merged.loc[:, 'atributo'] = df_merged['atributo'].replace(mapeo_lgbtq)
df_merged.loc[:, 'atributo'] = df_merged['atributo'].replace(mapeo_accesibilidad)


Volvamos a ver los atributos

categorias = df_merged['categoria_atributo'].unique()

# Iterar a través de cada categoría y mostrar los atributos únicos
for categoria in categorias:
    atributos = df_merged[df_merged['categoria_atributo'] == categoria]['atributo'].unique()
    print(f"{categoria}\n", atributos, "\n")

Vemos que dentro de offering se encuentra un atributo 'braille menu' esto es muy importante para determinar la accesibilidad asi que lo agruparemos en esa categoria. Ademas el atributo 'High chairs' tambien podria entrar dentro de la accesibilidad para problemas de altura.

# Cambiar la categoría a "Accessibility" para registros con "Offerings" y "Braille menu"
df_merged.loc[(df_merged['categoria_atributo'] == 'Offerings') & (df_merged['atributo'] == 'Braille menu'), 'categoria_atributo'] = 'Accessibility'
df_merged.loc[(df_merged['categoria_atributo'] == 'Amenities') & (df_merged['atributo'] == 'High chairs'), 'categoria_atributo'] = 'Accessibility'

# Muestra final
categorias = df_merged['categoria_atributo'].unique()

# Iterar a través de cada categoría y mostrar los atributos únicos
for categoria in categorias:
    atributos = df_merged[df_merged['categoria_atributo'] == categoria]['atributo'].unique()
    print(f"{categoria}\n", atributos, "\n")

Como paso final, almacenaré la tabla en un archivo parquet para su posterior carga a la base de datos.

df_merged.to_parquet(r"C:\Users\Juli\Desktop\Data Science\Archivos Proyecto Final\v0_misc.parquet")

# Muestra de la tabla final
df_merged

misc['id_misc'] = range(1, len(misc) + 1)

import pandas as pd 
misc = pd.read_parquet(r"C:\Users\mauri\OneDrive\Escritorio\proyectogrupal\data\v0_misc.parquet", engine='pyarrow')
misc.info()


# Lista de categorías que deseas excluir
categorias_excluir = ['Recycling', 'From the business','Payments','Dining options', 'Planning', 'Service options','Offerings']
atributos_excluir = ['Fireplace','Active military discounts','Quiz night', '3D movies','Swimming pool', 'Mechanic','Golf course', 'Stadium seating']

# Filtrar el DataFrame para quedarte solo con las filas que NO tengan esas categorías
filtrado = misc[~misc['categoria_atributo'].isin(categorias_excluir)]
filtrado = filtrado[~filtrado['atributo'].isin(atributos_excluir)]
filtrado.info()

# Muestra final
categorias = filtrado['categoria_atributo'].unique()

# Iterar a través de cada categoría y mostrar los atributos únicos
for categoria in categorias:
    atributos = filtrado[filtrado['categoria_atributo'] == categoria]['atributo'].unique()
    print(f"{categoria}\n", atributos, "\n")

filtrado.to_parquet("v1_misc.parquet")
