==> notebooks/ETL/Google/5.g_horarios.ipynb <==
En este notebook armaré una tabla que contenga los dias y horarios en los que abre cada negocio de los estados y categorías seleccionados.

import pandas as pd
import os

# Leemos la tabla v1.g_sitios que tenemos hasta el momento.
g_sitios = pd.read_parquet(r"C:\Users\Juli\Desktop\Data Science\Archivos Proyecto Final\v1.g_sitios.parquet", engine='pyarrow')
g_sitios.info()

g_sitios

Cambiaré el nombre de la columna "gmap_id" por "id_negocio".

g_sitios = g_sitios.rename(columns={'gmap_id': 'id_negocio'})

# Creamos la tabla horarios
g_horarios = g_sitios[['id_negocio','monday', 'tuesday', 'wednesday', 'thursday', 'friday', 'saturday','sunday']]
g_horarios

g_horarios

g_horarios.info()

Verificamos si hay nulos.

nulos = g_horarios.isnull().sum()
nulos

Guardamos el dataframe en un archivo parquet para luego cargar a nuestra base de datos.

g_horarios.to_parquet(r"C:\Users\Juli\Desktop\Data Science\Archivos Proyecto Final\v0_horarios.parquet")

import os
import pandas as pd
g_sitios = pd.read_parquet(r'C:\Users\mauri\OneDrive\Escritorio\proyectogrupal\data\v1_g_sitios.parquet')
# Eliminamos los dias de g_sitios
g_sitios = g_sitios.drop(columns=['monday', 'tuesday', 'wednesday', 'thursday', 'friday', 'saturday', 'sunday'])

# Guardamos el nuevo Dataframe.
os.makedirs('../../../data', exist_ok=True)
g_sitios.to_parquet('../../../data/v2.g_sitios.parquet', index=False)

# Muestra parcial de la tabla:
g_sitios.head()

# Guardamos el nuevo Dataframe g_horarios.
os.makedirs('../../../data', exist_ok=True)
g_horarios.to_parquet('../../../data/google_horarios.parquet', index=False)

# Muestra parcial de la tabla:
g_horarios.head()
