==> notebooks/ETL/Google/7.g_relativer_result.ipynb <==
En este notebook se trabajará con el dataset **"google_relative_results_sitios"** para su transformación. De resultado se obtendrá una tabla que contenga únicamente los id de los negocios que están en los cinco estados elegidos para este proyecto correspondientes a las categorías también preseleccionadas. La misma, luego, será subida a nuestra base de datos.
Este dataset contiene información de los id de los negocios y los id de los negocios relacionados.

Como primer paso voy a importar el dataset.

import pandas as pd

relative_results = pd.read_parquet(r"C:\Users\Juli\Desktop\Data Science\Archivos Proyecto Final\google_relative_results_sitios.parquet")
relative_results.head()

relative_results.info()

Importaré el dataframe "vi.g_sitios". Realizaré una unión entre ambos dataframes para luego filtrar por las categorías y estados elegidos. 

sitios= pd.read_parquet(r"C:\Users\Juli\Desktop\Data Science\Archivos Proyecto Final\v1.g_sitios.parquet")
sitios.head()

df_merged = pd.merge(relative_results, sitios, left_on='gmap_id', right_on='gmap_id', how='inner')
df_merged

Filtraré por las columnas de interés.

tabla_final = df_merged[['gmap_id', 'relative_results']]
tabla_final

Realizaré una búsqueda de **valores nulos.**

# Cuento los nulos en cada fila
cant_nulos = tabla_final.isnull().sum()
print(cant_nulos)

# Muestro las filas que contienen algún valor nulo
tabla_final[tabla_final.isnull().any(axis=1)]

# Calculo el porcentaje de nulos de la columna relative_results
nulos_relative_results = tabla_final['relative_results'].isnull().sum()
total_filas = len(tabla_final)
porcentaje_nulos = (nulos_relative_results / total_filas) * 100

print(porcentaje_nulos)

Podemos observar que la cantidad de nulos es baja y debido a que uno de los objetivos propuestos en este proyecto es identificar las mejores ubicaciones para nuevos locales basándose en la densidad de competidores y la demanda de usuarios, he decidido dejarlos. Son datos valiosos para el análisis. Voy a reemplazar los nulos por "sin competencia".

# Desactivo SettingWithCopyWarning
pd.options.mode.chained_assignment = None

# Reemplazo valores de tipo None o el texto "None" por "sin competencia"
tabla_final['relative_results'] = tabla_final['relative_results'].replace([None, "None"], "sin competencia")

# Filtro y muestro las filas donde relative_results es "sin competencia"
sin_competencia_filas = tabla_final[tabla_final['relative_results'] == "sin competencia"]
sin_competencia_filas

Voy a realizar una búsqueda de valores **duplicados.**

# Identifico y muestro filas duplicadas
duplicados = tabla_final[tabla_final.duplicated(keep=False)]
print(duplicados)

No se han encontrado filas duplicadas.

Revisaré si los tipos de datos de las columnas son los que necesito de acuerdo al modelo relacional creado para la base de datos.

print(tabla_final.dtypes)

# Modifico el nombre de la columna "gmap_id" por "id_negocio"
tabla_final = tabla_final.rename(columns={'gmap_id': 'id_negocio'})
tabla_final

Exportaré la tabla ya que está lista para ser utilizada en nuestra base de datos.

tabla_final.to_parquet(r"C:\Users\Juli\Desktop\v0_relative_results.parquet")
