==> notebooks/ETL/Google/8.g_reviews_5_estados.ipynb <==
import pandas as pd
import os

Lo que haremos aquí es recopilar los datos de las reviews de los estados que utilizaremos en el análisis. Los archivos en crudo se presentaron en formato Json, en archivos individuales para cada uno de los 51 estados de EEUU, por lo que procederemos a ordenar los mismos, en formato parquet, y por la dimension, en dataframes individuales, para luego crear una tabla única para los 5 estados pertinentes.

# New York

Se convierten los archivos Json a formato parquet

# Carpeta de entrada que contiene los archivos JSON
input_folder = r'C:\Users\debor\OneDrive\Documentos\Henry\PROYECTO FINAL local\Datasets\datasets_google\estados_json\NORESTE\review-New_York'
# Carpeta de salida para guardar los archivos Parquet
output_folder = r'C:\Users\debor\OneDrive\Documentos\Henry\PROYECTO FINAL local\Datasets\datasets_google\estados_parquet\NORESTE_parquet\Rev-New_York.parquet'

# Crear la carpeta de salida si no existe
os.makedirs(output_folder, exist_ok=True)

# Iterar sobre todos los archivos en la carpeta de entrada
for filename in os.listdir(input_folder):
    if filename.endswith('.json'):
        # Ruta completa del archivo JSON
        json_file_path = os.path.join(input_folder, filename)
        
        # Leer el archivo JSON
        try:
            df = pd.read_json(json_file_path, lines=True)
            
            # Crear nombre de archivo Parquet (sin la extensión .json)
            parquet_file_path = os.path.join(output_folder, filename.replace('.json', '.parquet'))
            
            # Guardar el DataFrame como archivo Parquet
            df.to_parquet(parquet_file_path, index=False, compression='snappy', engine='pyarrow')
            
            print(f"Convertido: {filename} a {parquet_file_path}")
        except Exception as e:
            print(f"Error al convertir {filename}: {e}")

print("Conversión completada.")

# Leer el archivo Parquet
Rev_New_York_1 = pd.read_parquet(r'C:\Users\debor\OneDrive\Documentos\Henry\PROYECTO FINAL local\Datasets\datasets_google\estados_parquet\NORESTE_parquet\Rev-New_York.parquet')

print(Rev_New_York_1.info())

Se concatenan los archivos parquet para formar un dataframe

# Carpeta que contiene los archivos Parquet
input_folder_parquet = r'C:\Users\debor\OneDrive\Documentos\Henry\PROYECTO FINAL local\Datasets\datasets_google\estados_parquet\NORESTE_parquet\Rev-New_York.parquet'

# Lista para almacenar DataFrames
df_list_New_York = []

# Iterar sobre todos los archivos Parquet en la carpeta de entrada
for filename in os.listdir(input_folder_parquet):
    if filename.endswith('.parquet'):
        # Ruta completa del archivo Parquet
        parquet_file_path_ny = os.path.join(input_folder_parquet, filename)
        
        # Leer el archivo Parquet y añadir el DataFrame a la lista
        df = pd.read_parquet(parquet_file_path_ny)
        df_list_New_York.append(df)

# Concatenar todos los DataFrames en uno solo
df_new_york = pd.concat(df_list_New_York, ignore_index=True)

# Verificar el DataFrame combinado
print(df_new_york.shape)
print

Se elimina columna irrelevante para el análisis

df_new_york = df_new_york.drop(columns=['pics'])

Se convierten las columnas necesarias

# Se convierte la columna 'user_id' a tipo object
df_new_york['user_id'] = df_new_york['user_id'].astype('str')
print(df_new_york['user_id'].head(3))

# Se convierte la columna 'time' a datetime y luego a solo fecha manteniendo tipo datetime
df_new_york['time'] = pd.to_datetime(df_new_york['time'], unit='ms').dt.normalize()
print(df_new_york['time'].head(3))

# Se normaliza la columna 'resp' para agregar su contenido como dos nuevas columnas a df_new_york
resp_df_ny = pd.json_normalize(df_new_york['resp'])
df_new_york[['resp_text', 'resp_time']] = resp_df_ny[['text', 'time']]

# Se convierte la columna 'resp_time' a datetime y luego a solo fecha manteniendo tipo datetime
df_new_york['resp_time'] = pd.to_datetime(df_new_york['resp_time'], unit='ms').dt.normalize()
print(df_new_york['resp_time'].head(3))

# Se elimina la columna 'resp' original
df_new_york = df_new_york.drop(columns=['resp'])

# Se renombra la columna gmap_id como 'id_negocio', para mayor comprensión en el análisis
df_new_york = df_new_york.rename(columns={'gmap_id': 'id_negocio'})

# Se reordena la columna 'id_negocio'
columnas = list(df_new_york.columns)
columnas.insert(0, columnas.pop(columnas.index('id_negocio')))
df_new_york = df_new_york[columnas]

# Se añade la columna 'review_id' autoincremental
df_new_york.insert(0, 'review_id', range(1, 1 + len(df_new_york)))

# Se añade la columna 'Estado'
df_new_york.insert(2, 'estado', 'New York')

# Se verifica como queda el dataframe
print(df_new_york.info())

# California

Se pasan de formato Json a Parquet

# Carpeta de entrada que contiene los archivos JSON
input_folder = r'C:\Users\debor\OneDrive\Documentos\Henry\PROYECTO FINAL local\Datasets\datasets_google\estados_json\OESTE\review-California'
# Carpeta de salida para guardar los archivos Parquet
output_folder = r'C:\Users\debor\OneDrive\Documentos\Henry\PROYECTO FINAL local\Datasets\datasets_google\estados_parquet\OESTE.parquet\Rev-California.parquet'

# Crear la carpeta de salida si no existe
os.makedirs(output_folder, exist_ok=True)

# Iterar sobre todos los archivos en la carpeta de entrada
for filename in os.listdir(input_folder):
    if filename.endswith('.json'):
        # Ruta completa del archivo JSON
        json_file_path = os.path.join(input_folder, filename)
        
        # Leer el archivo JSON
        try:
            df = pd.read_json(json_file_path, lines=True)
            
            # Crear nombre de archivo Parquet (sin la extensión .json)
            parquet_file_path = os.path.join(output_folder, filename.replace('.json', '.parquet'))
            
            # Guardar el DataFrame como archivo Parquet
            df.to_parquet(parquet_file_path, index=False, compression='snappy', engine='pyarrow')
            
            print(f"Convertido: {filename} a {parquet_file_path}")
        except Exception as e:
            print(f"Error al convertir {filename}: {e}")

print("Conversión completada.")

# Leer el archivo Parquet
Rev_California_1 = pd.read_parquet(r'C:\Users\debor\OneDrive\Documentos\Henry\PROYECTO FINAL local\Datasets\datasets_google\estados_parquet\OESTE.parquet\Rev-New_Mexico.parquet')

# Mostrar las primeras filas
print(Rev_California_1.info())

Se concatenan los archivos .parquet para formar un Dataframe

# Carpeta que contiene los archivos Parquet
input_folder_parquet = r'C:\Users\debor\OneDrive\Documentos\Henry\PROYECTO FINAL local\Datasets\datasets_google\estados_parquet\OESTE.parquet\Rev-California.parquet'

# Lista para almacenar DataFrames
df_list_California = []

# Iterar sobre todos los archivos Parquet en la carpeta de entrada
for filename in os.listdir(input_folder_parquet):
    if filename.endswith('.parquet'):
        # Ruta completa del archivo Parquet
        parquet_file_path_nmex = os.path.join(input_folder_parquet, filename)
        
        # Leer el archivo Parquet y añadir el DataFrame a la lista
        df = pd.read_parquet(parquet_file_path_nmex)
        df_list_California.append(df)

# Concatenar todos los DataFrames en uno solo
df_california = pd.concat(df_list_California, ignore_index=True)

# Verificar el DataFrame combinado
print(df_california.shape)
print

Se elimina columna irrelevante para el análisis

df_california = df_california.drop(columns=['pics'])

Se convierten las columnas necesarias

# Se convierte la columna 'user_id' a tipo object
df_california['user_id'] = df_california['user_id'].astype('str')
print(df_california['user_id'].head(3))


# Se convierte la columna 'time' a datetime y luego a solo fecha manteniendo tipo datetime
df_california['time'] = pd.to_datetime(df_california['time'], unit='ms').dt.normalize()
print(df_california['time'].head(3))

# Se normaliza la columna 'resp' para agregar su contenido como dos nuevas columnas a df_new_york
resp_df_cal = pd.json_normalize(df_california['resp'])
df_california[['resp_text', 'resp_time']] = resp_df_cal[['text', 'time']]

# Se convierte la columna 'time' a datetime y luego a solo fecha manteniendo tipo datetime
df_california['resp_time'] = pd.to_datetime(df_california['resp_time'], unit='ms').dt.normalize()
print(df_california['resp_time'].head(3))

# Se elimina la columna 'resp' original
df_california = df_california.drop(columns=['resp'])

# Se renombra la columna gmap_id como 'id_negocio', para mayor comprensión en el análisis
df_california = df_california.rename(columns={'gmap_id': 'id_negocio'})

# Se reordena la columna 'id_negocio'
columnas = list(df_california.columns)
columnas.insert(0, columnas.pop(columnas.index('id_negocio')))
df_california = df_california[columnas]

# Se añade la columna 'review_id' autoincremental desde 2700001 para continuar desde el df_new_york
df_california.insert(0, 'review_id', range(2700001, 2700001 + len(df_california)))

# Se añade la columna 'Estado'
df_california.insert(2, 'estado', 'California')

# Se verifica como queda el dataframe
print(df_california.info())

# Pennsylvania

Se convierten los archivos Json a formato parquet

# Carpeta de entrada que contiene los archivos JSON
input_folder = r'C:\Users\debor\OneDrive\Documentos\Henry\PROYECTO FINAL local\Datasets\datasets_google\estados_json\NORESTE\review-Pennsylvania'
# Carpeta de salida para guardar los archivos Parquet
output_folder = r'C:\Users\debor\OneDrive\Documentos\Henry\PROYECTO FINAL local\Datasets\datasets_google\estados_parquet\NORESTE_parquet\Rev-Pennsylvania.parquet'

# Crear la carpeta de salida si no existe
os.makedirs(output_folder, exist_ok=True)

# Iterar sobre todos los archivos en la carpeta de entrada
for filename in os.listdir(input_folder):
    if filename.endswith('.json'):
        # Ruta completa del archivo JSON
        json_file_path = os.path.join(input_folder, filename)
        
        # Leer el archivo JSON
        try:
            df = pd.read_json(json_file_path, lines=True)
            
            # Crear nombre de archivo Parquet (sin la extensión .json)
            parquet_file_path = os.path.join(output_folder, filename.replace('.json', '.parquet'))
            
            # Guardar el DataFrame como archivo Parquet
            df.to_parquet(parquet_file_path, index=False, compression='snappy', engine='pyarrow')
            
            print(f"Convertido: {filename} a {parquet_file_path}")
        except Exception as e:
            print(f"Error al convertir {filename}: {e}")

print("Conversión completada.")

# Leer el archivo Parquet
Rev_Pennsylvania_1 = pd.read_parquet(r'C:\Users\debor\OneDrive\Documentos\Henry\PROYECTO FINAL local\Datasets\datasets_google\estados_parquet\NORESTE_parquet\Rev-Pennsylvania.parquet')

print(Rev_Pennsylvania_1.info())

Se concatenan los archivos parquet para formar un solo dataframe

# Carpeta que contiene los archivos Parquet
input_folder_parquet = r'C:\Users\debor\OneDrive\Documentos\Henry\PROYECTO FINAL local\Datasets\datasets_google\estados_parquet\NORESTE_parquet\Rev-Pennsylvania.parquet'

# Lista para almacenar DataFrames
df_list_Pennsylvania = []

# Iterar sobre todos los archivos Parquet en la carpeta de entrada
for filename in os.listdir(input_folder_parquet):
    if filename.endswith('.parquet'):
        # Ruta completa del archivo Parquet
        parquet_file_path_psv = os.path.join(input_folder_parquet, filename)
        
        # Leer el archivo Parquet y añadir el DataFrame a la lista
        df = pd.read_parquet(parquet_file_path_psv)
        df_list_Pennsylvania.append(df)

# Concatenar todos los DataFrames en uno solo
df_pennsylvania = pd.concat(df_list_Pennsylvania, ignore_index=True)

# Verificar el DataFrame combinado
print(df_pennsylvania.shape)
print

Se elimina columna irrelevante para el análisis

df_pennsylvania = df_pennsylvania.drop(columns=['pics'])

Se convierten las columnas necesarias

# Se convierte la columna 'user_id' a tipo object
df_pennsylvania['user_id'] = df_pennsylvania['user_id'].astype('str')
print(df_pennsylvania['user_id'].head(3))

# Se convierte la columna 'time' a datetime y luego a solo fecha manteniendo tipo datetime
df_pennsylvania['time'] = pd.to_datetime(df_pennsylvania['time'], unit='ms').dt.normalize()
print(df_pennsylvania['time'].head(3))

# Se normaliza la columna 'resp' para agregar su contenido como dos nuevas columnas a df_new_york
resp_df_pen = pd.json_normalize(df_pennsylvania['resp'])
df_pennsylvania[['resp_text', 'resp_time']] = resp_df_pen[['text', 'time']]

# Se convierte la columna 'resp_time' a datetime y luego a solo fecha manteniendo tipo datetime
df_pennsylvania['resp_time'] = pd.to_datetime(df_pennsylvania['resp_time'], unit='ms').dt.normalize()
print(df_pennsylvania['resp_time'].head(3))

# Se elimina la columna 'resp' original
df_pennsylvania = df_pennsylvania.drop(columns=['resp'])

# Se renombra la columna gmap_id como 'id_negocio', para mayor comprensión en el análisis
df_pennsylvania = df_pennsylvania.rename(columns={'gmap_id': 'id_negocio'})

# Se reordena la columna 'id_negocio'
columnas = list(df_pennsylvania.columns)
columnas.insert(0, columnas.pop(columnas.index('id_negocio')))
df_pennsylvania = df_pennsylvania[columnas]

# Se añade la columna 'review_id' autoincremental desde 4500001 para continuar desde el df_california
df_pennsylvania.insert(0, 'review_id', range(4500001, 4500001 + len(df_pennsylvania)))

# Se añade la columna 'Estado'
df_pennsylvania.insert(2, 'estado', 'Pennsylvania')

# Se verifica como queda el dataframe
print(df_pennsylvania.info())

# Florida

Se convierten los archivos Json a formato parquet

# Carpeta de entrada que contiene los archivos JSON
input_folder = r'C:\Users\debor\OneDrive\Documentos\Henry\PROYECTO FINAL local\Datasets\datasets_google\estados_json\SUR\review-Florida'
# Carpeta de salida para guardar los archivos Parquet
output_folder = r'C:\Users\debor\OneDrive\Documentos\Henry\PROYECTO FINAL local\Datasets\datasets_google\estados_parquet\SUR_parquet\Rev-Florida.parquet'

# Crear la carpeta de salida si no existe
os.makedirs(output_folder, exist_ok=True)

# Iterar sobre todos los archivos en la carpeta de entrada
for filename in os.listdir(input_folder):
    if filename.endswith('.json'):
        # Ruta completa del archivo JSON
        json_file_path = os.path.join(input_folder, filename)
        
        # Leer el archivo JSON
        try:
            df = pd.read_json(json_file_path, lines=True)
            
            # Crear nombre de archivo Parquet (sin la extensión .json)
            parquet_file_path = os.path.join(output_folder, filename.replace('.json', '.parquet'))
            
            # Guardar el DataFrame como archivo Parquet
            df.to_parquet(parquet_file_path, index=False, compression='snappy', engine='pyarrow')
            
            print(f"Convertido: {filename} a {parquet_file_path}")
        except Exception as e:
            print(f"Error al convertir {filename}: {e}")

print("Conversión completada.")

# Leer el archivo Parquet
Rev_Florida_1 = pd.read_parquet(r'C:\Users\debor\OneDrive\Documentos\Henry\PROYECTO FINAL local\Datasets\datasets_google\estados_parquet\SUR_parquet\Rev-Florida.parquet')

print(Rev_Florida_1.info())

Se concatenan los archivos parquet para formar un solo dataframe

# Carpeta que contiene los archivos Parquet
input_folder_parquet = r'C:\Users\debor\OneDrive\Documentos\Henry\PROYECTO FINAL local\Datasets\datasets_google\estados_parquet\SUR_parquet\Rev-Florida.parquet'

# Lista para almacenar DataFrames
df_list_Florida = []

# Iterar sobre todos los archivos Parquet en la carpeta de entrada
for filename in os.listdir(input_folder_parquet):
    if filename.endswith('.parquet'):
        # Ruta completa del archivo Parquet
        parquet_file_path_psv = os.path.join(input_folder_parquet, filename)
        
        # Leer el archivo Parquet y añadir el DataFrame a la lista
        df = pd.read_parquet(parquet_file_path_psv)
        df_list_Florida.append(df)

# Concatenar todos los DataFrames en uno solo
df_florida = pd.concat(df_list_Florida, ignore_index=True)

# Verificar el DataFrame combinado
print(df_florida.shape)
print

Se elimina columna irrelevante para el análisis

df_florida = df_florida.drop(columns=['pics'])

Se convierten las columnas necesarias

# Se convierte la columna 'user_id' a tipo object
df_florida['user_id'] = df_florida['user_id'].astype('str')
print(df_florida['user_id'].head(3))

# Se convierte la columna 'time' a datetime y luego a solo fecha manteniendo tipo datetime
df_florida['time'] = pd.to_datetime(df_florida['time'], unit='ms').dt.normalize()
print(df_florida['time'].head(3))

# Se normaliza la columna 'resp' para agregar su contenido como dos nuevas columnas a df_new_york
resp_df_fl = pd.json_normalize(df_florida['resp'])
df_florida[['resp_text', 'resp_time']] = resp_df_fl[['text', 'time']]

# Se convierte la columna 'resp_time' a datetime y luego a solo fecha manteniendo tipo datetime
df_florida['resp_time'] = pd.to_datetime(df_florida['resp_time'], unit='ms').dt.normalize()
print(df_florida['resp_time'].head(3))

# Se elimina la columna 'resp' original
df_florida = df_florida.drop(columns=['resp'])

# Se renombra la columna gmap_id como 'id_negocio', para mayor comprensión en el análisis
df_florida = df_florida.rename(columns={'gmap_id': 'id_negocio'})

# Se reordena la columna 'id_negocio'
columnas = list(df_florida.columns)
columnas.insert(0, columnas.pop(columnas.index('id_negocio')))
df_florida = df_florida[columnas]

# Se añade la columna 'review_id' autoincremental desde 6900001 para continuar desde el df_pennsylvania
df_florida.insert(0, 'review_id', range(6900001, 6900001 + len(df_florida)))

# Se añade la columna 'estado'
df_florida.insert(2, 'estado', 'Pennsylvania')

# Se verifica como queda el dataframe
print(df_florida.info())

df_florida.tail()

# Texas

Se convierten los archivos Json a formato parquet

# Carpeta de entrada que contiene los archivos JSON
input_folder = r'C:\Users\debor\OneDrive\Documentos\Henry\PROYECTO FINAL local\Datasets\datasets_google\estados_json\SUR\review-Texas'
# Carpeta de salida para guardar los archivos Parquet
output_folder = r'C:\Users\debor\OneDrive\Documentos\Henry\PROYECTO FINAL local\Datasets\datasets_google\estados_parquet\SUR_parquet\Rev-Texas.parquet'

# Crear la carpeta de salida si no existe
os.makedirs(output_folder, exist_ok=True)

# Iterar sobre todos los archivos en la carpeta de entrada
for filename in os.listdir(input_folder):
    if filename.endswith('.json'):
        # Ruta completa del archivo JSON
        json_file_path = os.path.join(input_folder, filename)
        
        # Leer el archivo JSON
        try:
            df = pd.read_json(json_file_path, lines=True)
            
            # Crear nombre de archivo Parquet (sin la extensión .json)
            parquet_file_path = os.path.join(output_folder, filename.replace('.json', '.parquet'))
            
            # Guardar el DataFrame como archivo Parquet
            df.to_parquet(parquet_file_path, index=False, compression='snappy', engine='pyarrow')
            
            print(f"Convertido: {filename} a {parquet_file_path}")
        except Exception as e:
            print(f"Error al convertir {filename}: {e}")

print("Conversión completada.")

# Leer el archivo Parquet
Rev_Texas_1 = pd.read_parquet(r'C:\Users\debor\OneDrive\Documentos\Henry\PROYECTO FINAL local\Datasets\datasets_google\estados_parquet\SUR_parquet\Rev-Texas.parquet')

print(Rev_Texas_1.info())

Se concatenan los archivos parquet para formar un solo dataframe

# Carpeta que contiene los archivos Parquet
input_folder_parquet = r'C:\Users\debor\OneDrive\Documentos\Henry\PROYECTO FINAL local\Datasets\datasets_google\estados_parquet\NORESTE_parquet\Rev-Pennsylvania.parquet'

# Lista para almacenar DataFrames
df_list_Texas = []

# Iterar sobre todos los archivos Parquet en la carpeta de entrada
for filename in os.listdir(input_folder_parquet):
    if filename.endswith('.parquet'):
        # Ruta completa del archivo Parquet
        parquet_file_path_psv = os.path.join(input_folder_parquet, filename)
        
        # Leer el archivo Parquet y añadir el DataFrame a la lista
        df = pd.read_parquet(parquet_file_path_psv)
        df_list_Texas.append(df)

# Concatenar todos los DataFrames en uno solo
df_texas = pd.concat(df_list_Texas, ignore_index=True)

# Verificar el DataFrame combinado
print(df_texas.shape)
print

Se elimina columna irrelevante para el análisis

df_texas = df_texas.drop(columns=['pics'])

Se convierten las columnas necesarias

# Se convierte la columna 'user_id' a tipo object
df_texas['user_id'] = df_texas['user_id'].astype('str')
print(df_texas['user_id'].head(3))

# Se convierte la columna 'time' a datetime y luego a solo fecha manteniendo tipo datetime
df_texas['time'] = pd.to_datetime(df_texas['time'], unit='ms').dt.normalize()
print(df_texas['time'].head(3))

# Se normaliza la columna 'resp' para agregar su contenido como dos nuevas columnas a df_new_york
resp_df_tx = pd.json_normalize(df_texas['resp'])
df_texas[['resp_text', 'resp_time']] = resp_df_tx[['text', 'time']]

# Se convierte la columna 'resp_time' a datetime y luego a solo fecha manteniendo tipo datetime
df_texas['resp_time'] = pd.to_datetime(df_texas['resp_time'], unit='ms').dt.normalize()
print(df_texas['resp_time'].head(3))

# Se elimina la columna 'resp' original
df_texas = df_texas.drop(columns=['resp'])

# Se renombra la columna gmap_id como 'id_negocio', para mayor comprensión en el análisis
df_texas = df_texas.rename(columns={'gmap_id': 'id_negocio'})

# Se reordena la columna 'id_negocio'
columnas = list(df_texas.columns)
columnas.insert(0, columnas.pop(columnas.index('id_negocio')))
df_texas = df_texas[columnas]

# Se añade la columna 'review_id' autoincremental desde 9750001 para continuar desde el df_florida
df_texas.insert(0, 'review_id', range(9750001, 9750001 + len(df_texas)))

# Se añade la columna 'estado'
df_texas.insert(2, 'estado', 'Pennsylvania')

# Se verifica como queda el dataframe
print(df_texas.info())

# Se concatenan los  DataFrames anteriores para formar uno solo
g_estados_seleccionados = pd.concat([df_new_york, df_california, df_pennsylvania, df_florida, df_texas])

g_estados_seleccionados_2 = g_estados_seleccionados.to_parquet(r'C:\Users\debor\OneDrive\Documentos\Henry\PROYECTO FINAL local\Datasets\datasets_google\g_estados_seleccionados_2.parquet', engine='auto', compression='snappy')

print(g_estados_seleccionados.info())

nulos_por_columna = g_estados_seleccionados.isnull().sum()
print(nulos_por_columna)

g_sitios = pd.read_parquet(r'C:\Users\debor\OneDrive\Documentos\Henry\PROYECTO FINAL local\Datasets\datasets_google\g_sitios_v3.parquet')

# Obtenemos los ID unicos
g_sitios_ids = g_sitios['id_negocio'].unique()

# Filtrar las tablas por los IDs que coincidan con los de g_sitios
g_estados_seleccionados = g_estados_seleccionados[g_estados_seleccionados['id_negocio'].isin(g_sitios_ids)]

# Resetear los índices del DataFrame filtrado
g_estados_seleccionados = g_estados_seleccionados.reset_index(drop=True)

# Crear una nueva columna 'review_id' con valores consecutivos solo para el DataFrame filtrado
g_estados_seleccionados['review_id'] = range(1, len(g_estados_seleccionados) + 1)

g_estados_seleccionados.info()

g_estados_seleccionados.to_parquet(r'C:\Users\debor\OneDrive\Documentos\Henry\PROYECTO FINAL local\Datasets\datasets_google\g_estados_seleccionados.parquet', engine='auto', compression='snappy')
